+++ stack.h
@@ -3,13 +3,13 @@
 */
 #define STACK_MAX 1000
 struct Stack {
-    int     data[STACK_MAX];
-    int     size;
+    long int     data[STACK_MAX];
+    unsigned int     size;
 };
 
 typedef struct Stack Stack;
 
 void Stack_Init(Stack *S);
 int Stack_Top(Stack *S);
-void Stack_Push(Stack *S, int d);
+void Stack_Push(Stack *S, long int d);
 void Stack_Pop(Stack *S);
